# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "United Tractor API",
        "description": "United Tractor API documentation",
        "contact": {
            "email": "fajarshidik78@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/category-products": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "View all category and their product",
                "operationId": "91be97e267a9a2583036f32363b690cc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "furniture"
                                                    },
                                                    "product": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "category_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "kursi"
                                                                },
                                                                "price": {
                                                                    "type": "integer",
                                                                    "example": 100000
                                                                },
                                                                "image": {
                                                                    "type": "string",
                                                                    "example": "IMG-0001"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Make a new category Product",
                "operationId": "693a50d49e44b321d1da31ac1893d06a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "furniture"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "furniture"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/category-products/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "View category and their product by id",
                "operationId": "4e5f2fa529e30f8642b385a2a6c74ba5",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "furniture"
                                                },
                                                "product": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "category_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "kursi"
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "example": 100000
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "example": "IMG-0001"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a category Product",
                "operationId": "32e2ac4c1bb52263779c7a05ac62fcda",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a category Product",
                "operationId": "d4df2578079b16e5a26d1b107680a255",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "equipment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "equipment"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerToken": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer &lt;token&gt;)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}